Finished the application

+++ src/main/java/KthDatabaseApp/model/TeachingActivityDTO.java
@@ -2,5 +2,5 @@ package KthDatabaseApp.model;
 
 public interface TeachingActivityDTO {
     public String getActivityName();
-    public int getMultiplicationFactor();
+    public double getMultiplicationFactor();
 }
